task exec update
structures moved
